+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -86,6 +86,7 @@ public let masterRuleList = RuleList(rules:
     EmptyParametersRule.self,
     EmptyParenthesesWithTrailingClosureRule.self,
     ExplicitInitRule.self,
+    ExpressibleByArrayLiteralInitRule.self,
     FileHeaderRule.self,
     FileLengthRule.self,
     FirstWhereRule.self,