+++ CHANGELOG.md
@@ -73,6 +73,13 @@
   [JP Simard](https://github.com/jpsim)
   [#1077](https://github.com/realm/SwiftLint/issues/1077)
 
+* Add `expressible_by_array_literal_init` rule that flags usage of the
+  initializer declared in `ExpressibleByArrayLiteral` which shouldn't be
+  called directly. Instead, you should use an array literal anywhere a
+  concrete type conforming to the protocol is expected by the context.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#1096](https://github.com/realm/SwiftLint/issues/1096)
+
 ##### Bug Fixes
 
 * Ignore close parentheses on `vertical_parameter_alignment` rule.